backend/requirements.txt
DJango>=3.0.6,<3.1
wheel>=0.36.2,<.37
djangorestframework>=3.12.2,<3.13.0
psycopg2-binary==2.8.6
dj-rest-auth>=2.1,<2.2
djangorestframework-simplejwt>=4.6.0,<4.7
django-allauth>=0.44.0,<0.45
django-filter>=2.4.0, <2.5
django-rest-auth>=0.9.5, <1.0
channels>=3.0.3, <3.1
channels_redis>=3.2.0, <3.3
django-cors-headers>=3.7.0, <3.8